뷰 2개로 분리
